requirements.txt
numpy
pandas
psycopg2-binary
pyarrow
SQLAlchemy
openpyxl